+++ src/main/java/org/kish/KishServer.java
@@ -1,6 +1,6 @@
 package org.kish;
 
-import org.kish.dataBase.ExpandedDataBase;
+import org.kish.dataBase.Config;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.event.EventListener;
@@ -11,29 +11,21 @@ import java.util.HashMap;
 
 @SpringBootApplication
 public class KishServer {
-    public static ExpandedDataBase mainDataBase = null;
-    public static ExpandedDataBase mainSettings = null;
+    public static Config CONFIG = null;
     public static FirebaseManager firebaseManager = null;
 
     public static void main(String[] args) {
-        mainDataBase = new ExpandedDataBase("db/kish_main_db.json");
-        mainSettings = new ExpandedDataBase("db/kish2020.json");
-        mainSettings.setSaveWithPrettyGson(true);
+        // TODO : DB ID, DB PW, DB HOST, DB PORT, FB messaging_tokens, FB db url, FB service account
+        CONFIG = new Config("db/config.json");
+        CONFIG.setSaveWithPrettyGson(true);
         firebaseManager = new FirebaseManager();
         SpringApplication.run(KishServer.class, args);
 
         firebaseManager.sendFCMToAdmin("Server started", "서버가 시작되었습니다.\n" + new Date().toString(), new HashMap<>());
     }
 
-    /**
-     * @return 기본 데이터베이스
-     */
-    public ExpandedDataBase getMainDataBase(){
-        return mainDataBase;
-    }
-
-    public ExpandedDataBase getMainSettings(){
-        return mainSettings;
+    public Config getMainSettings(){
+        return CONFIG;
     }
 
     public FirebaseManager getFirebaseManager(){
